+++ src/gromacs/utility/smalloc.cpp
@@ -79,7 +79,7 @@ void *save_malloc(const char *name, const char *file, int line, size_t size)
             gmx_fatal(errno, __FILE__, __LINE__,
                       "Not enough memory. Failed to malloc %" PRId64 " bytes for %s\n"
                       "(called from file %s, line %d)",
-                      (int64_t)size, name, file, line);
+                      static_cast<int64_t>(size), name, file, line);
         }
         (void) memset(p, 0, size);
     }
@@ -120,13 +120,13 @@ void *save_calloc(const char *name, const char *file, int line,
         }
         memset(p, 0, (size_t) (nelem * elsize));
 #else
-        if ((p = calloc((size_t)nelem, (size_t)elsize)) == nullptr)
+        if ((p = calloc(nelem, elsize)) == nullptr)
         {
             gmx_fatal(errno, __FILE__, __LINE__,
                       "Not enough memory. Failed to calloc %" PRId64
                       " elements of size %" PRId64
                       " for %s\n(called from file %s, line %d)",
-                      (int64_t)nelem, (int64_t)elsize, name, file, line);
+                      static_cast<int64_t>(nelem), static_cast<int64_t>(elsize), name, file, line);
         }
 #endif
     }
@@ -156,18 +156,18 @@ void *save_realloc(const char *name, const char *file, int line, void *ptr,
 #endif
         if (ptr == nullptr)
         {
-            p = malloc((size_t)size);
+            p = malloc(size);
         }
         else
         {
-            p = realloc(ptr, (size_t)size);
+            p = realloc(ptr, size);
         }
         if (p == nullptr)
         {
             gmx_fatal(errno, __FILE__, __LINE__,
                       "Not enough memory. Failed to realloc %" PRId64 " bytes for %s, %s=%x\n"
                       "(called from file %s, line %d)",
-                      (int64_t)size, name, name, ptr, file, line);
+                      static_cast<int64_t>(size), name, name, ptr, file, line);
         }
     }
     return p;
@@ -237,7 +237,7 @@ void *save_calloc_aligned(const char *name, const char *file, int line,
     void *aligned = save_malloc_aligned(name, file, line, nelem, elsize, alignment);
     if (aligned != nullptr)
     {
-        memset(aligned, 0, (size_t)(nelem * elsize));
+        memset(aligned, 0, static_cast<size_t>(nelem * elsize));
     }
     return aligned;
 }
